Add files via upload
run_analysis.R is an R Script containing a function called run_analysis(), that (1) merges the training and the test sets from the original study, to create one single set; (2) extracts only the measurements on the mean and standard deviation for each measurement; (3) uses descriptive activity names to name the activities in the data set; (4) appropriately labels the data set with descriptive variable names; (5) from the data set in step 4, creates a second, independent tidy data set with the average of each variable for each activity and each subject. 

Follow the instructions to make the script work in your computer: 
(1) Save the UCI HAR Dataset in your current working directory. 
(2) Copy and paste the script's content to your R console and press ENTER, in order to store the function in your global environment. 
(3) Call the function run_analysis(). It will create two txt files (totaldata and average) containing the above mentioned datasets, and store these two datasets in homonimous variables.